// angular-src/src/models/Plate.ts
/*
 * @Description: Model for Plate
 * @Author: Guozhi Tang
 * @Date: 2019-07-24 10:50:13
 * @Github: https://github.com/GuozhiTang/Bio-WebApp
 * @LastEditors: Guozhi Tang
 * @LastEditTime: 2019-08-30 14:44:18
 */
export class Plate {
  id: Number;
  barcode: Number;
  name: String;
  coor: String;
  volume: String;
  description: String;
}